Simplify storageInfo metrics
