+++ collector/helper.go
@@ -103,6 +103,34 @@ func (e *Exporter) removeCommas(str string) (float64, error) {
 	return convertedStr, nil
 }
 
+func (e *Exporter) bytesConverter(str string) (float64, error) {
+	type errorString struct {
+		s string
+	}
+	var bytesValue float64
+	level.Debug(e.logger).Log("msg", "Converting size to bytes")
+	num, err := e.removeCommas(str)
+	if err != nil {
+		return 0, err
+	}
+
+	if strings.Contains(str, "bytes") {
+		bytesValue = num
+	} else if strings.Contains(str, "KB") {
+		bytesValue = num * 1024
+	} else if strings.Contains(str, "MB") {
+		bytesValue = num * 1024 * 1024
+	} else if strings.Contains(str, "GB") {
+		bytesValue = num * 1024 * 1024 * 1024
+	} else if strings.Contains(str, "TB") {
+		bytesValue = num * 1024 * 1024 * 1024 * 1024
+	} else {
+		return 0, fmt.Errorf("Could not convert %s to bytes", str)
+	}
+	level.Debug(e.logger).Log("msg", "Successfully converted string to bytes", "string", str, "value", bytesValue)
+	return bytesValue, nil
+}
+
 func b2f(b bool) float64 {
 	if b {
 		return 1
